Re-do project properties hierarchy
Add Project Type, Date Picker for Project Date, and use more general terms for Author (vs. Designer) and Location (vs. Venue), with overrides provided.

+++ src/main/java/com/mhschmieder/fxguitoolkit/layout/LayoutFactory.java
@@ -287,8 +287,8 @@ public final class LayoutFactory {
      */
     public static GridPane makeGridPane( final Pos alignment,
                                          final Insets padding,
-                                         final int hgap,
-                                         final int vgap ) {
+                                         final double hgap,
+                                         final double vgap ) {
         final GridPane gridPane = new GridPane();
 
         gridPane.setAlignment( alignment );
